Created a dll for both C++ and C# usage
Created a second project called SDLWorld, which contains all the logic to render a world on screen using SDL
Moved Color struct and SDLWrapper class from the ProcedularWorlds project to the SDLWorld project
Removed ImGui
Renamed Generator class to HeightmapGenerator class
Removed PerlinGenerator class
Added Generator class, the main file for the whole library. The user will be able to plug in different generators into the main one (e.g. a heightmap generator, a river generator, a village generator, etc.)
Added dll syntax where appropriate so the library can be exported as a dll to be used in both C++ and C#